Big update to new freight classification
Changing the technology classification of FTT-Freight to add vans and three-wheelers. Cost data and shares data is dummy.

This version successfully exports CSVs and the model runs. The shares equation is broken but learning and cumulative capacities appear to function as expected.

Some variables have been removed from the FTT-Freight Masterfile and  'VariableListing.csv' because they were not used. Some elements of the freight cost matrix have also been removed for the same reason.

Also some general formatting of the freight Masterfile.